requirements.txt
Django
djangorestframework
django-cors-headers
psycopg
channels
channels-redis
daphne
Pillow
django-websockets